company-file-analyzer/requirements.txt
# Company File Analyzer - Python Dependencies

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
jinja2>=3.1.2

# Queue
redis>=5.0.0
rq>=1.16.0

# Document processing
pdfplumber>=0.10.0
python-docx>=1.1.0
PyPDF2>=3.0.0

# Image/OCR
Pillow>=10.0.0
pytesseract>=0.3.10

# HTTP client
requests>=2.31.0

# Utils
python-dotenv>=1.0.0
